backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
qrcode==7.4.2
pillow==10.1.0
supabase==1.0.4
pandas==2.1.2
openpyxl==3.1.2
python-dotenv==1.0.0
httpx==0.25.1
